feat: Adds an extra voucher option while creating invoices and code refactor
Extra Voucher option:
- Usually when automatic vouchers are created with payment modes as cash or Bank,
the transaction is tracked directly between mode of
payment a/c and the sale/purchase a/c.
- Only when the payment was in credit, the customer/supplier a/c was used.
- In order to track transactions using, customer/supplier
accounts irrespective of the mode of payment, a mode called "party" has been added.
- When the transaction is made with a party mode, A credit voucher will first be created between
sale/purchase a/c and the customer/supplier a/c.
And then a receipt voucher is made between the customer/supplier a/c and the mode of payment account (cash/bank).
- The existing modes of payment codes, 2 (bank), 3(cash) will behave the way they used to.
The codes 4 (bank - party) and 5 (cash - party) will use the new vouchering process of including the customer/supplier.

Refactor:
- Have reduced repetetion of code using variables to store conditional states.