feat: add view to handle Exceptions
This will add view to handle unhanlded exceptions from other view.
This is a temporarty implementation to avoid having to add extra
code for each API to get exception error logss (cause current in
setup everything is under try except) and also helps to move database
connections to use context managers. A proper solution has to be
developed in long term.

The response for Exception would be 200 status with
{"gkstatus": 2}, if we are using development.ini, it will also
send the error message.