fix: add service income to Profit and Loss gross profit
Service income was lost in calculation for Profit and Loss report.
As solution service invoice income is seperately calculated and
added to gross profit.

This is a workaround, and this should work because there is no
quantity wise sale/purchase/return supported by GNUKhata now and
credit note entries for services will be reflected only in "Discount
Paid" or "Discount Recieved", not in the "Sales A/C"