feat: add function to find Dr entry sign for a account head
This function checks if the increase in account head amounts to a
Debit entry. If it is an asset account, it will return +1 since a
Debit entry for asset means increase. For liability it will be -1
since a Debit entry for liability means decrease. The value for for
dr_sign will switch between 1 and -1 wrt the behaviour of the account
head to have Dr/Cr entry on increase or decrease.

This may not make any sense accounting wise, but is useful to
present Dr/Cr data on single column and also helps to calculate sums.