fix: remove blanket deletion of accounts with product name string in account name
When a item was deleted, all accounts with item name in the account
name string were being deleted. Now only accounts with name
`<item name> Purchase` and `<item name> Sale` are being deleted.

This is not an ideal solution, this is a temporary solution until
item table is connected to respective accounts via foreignkey. Also,
since there is a validation check setup to avoid creating products
if there are accounts with above name formats are present, so deleting
not connected account case is being handled.