refactor: simplify query to fetch godown id from stock table
The `fetchall` method of SQLAlchemy was used to query a single field
from a table, this is unnecessary since we have to parse the first
column and then the first row from it. This is fixed by using the
`scalar` method which will directly give us the value we need without
further parsing.