refactor: update migrate function
This will,

1. Merge `load_initial_data()` from `initdb.py` with `migrate()`
from `db_migrate.py`
2. Split the migrations and grouped under seperate context managers
to handle exceptions.
3. Update the database initialisation command.
4. Fix Wrong groupings. For example, delchal/invoice tables are
updated under a if condition that checks if log table exists.

The purpose of the function was to handle data migrations. But
data migrations can happen in between schema changes. This makes it
not practical to have a data migration initially or having to deal
with migrations at initialising and migrations at schema change.
Currently they are in a mixed form, so both functions are merged.
In future a proper tool should be used for database migrations.